refactor(pipeline): organize packaging into a folder with shared modules
Move the packagers into packages/pipeline/src/packaging/ (web, webpub,
epub, strip-runtime-bundle) so each output format has its own file —
webpub was previously buried in package-web.ts while epub already had
its own.

Extract the format-agnostic helpers out of the packager so nothing
outside packaging/ reaches into it for utilities:
- escapeHtml/escapeAttr/escapeInlineScriptJson -> html-escape.ts
- promoteFirstHeadingToH1/normalizeSectionRoles/htmlToXhtml -> html-semantics.ts
- buildTailwindCss/buildPreviewTailwindCss -> tailwind.ts (also restores
  the correct PIPELINE_PACKAGE_DIR resolution after the move)

Also emits toc/pageList/landmarks navigation collections in the webpub
manifest so a reader can build its own navigation UI.
