+++ packages/pipeline/src/packaging/strip-runtime-bundle.ts
