+++ packages/pipeline/src/pipeline-dag.ts
@@ -57,7 +57,7 @@ import {
   generateSpeechFile,
   type ProviderRouting,
 } from "./speech.js"
-import { packageAdtWeb } from "./package-web.js"
+import { packageAdtWeb } from "./packaging/web.js"
 import { processFixedLayoutPages, isFixedLayoutBook } from "./fixed-layout-rendering.js"
 import { runAccessibilityAssessment } from "./accessibility-assessment.js"
 import { loadBookConfig } from "./config.js"
